Update client sources structure
